server/model/iterators/CyclicalIterator: always start on same index
When creating a CyclicalIterator with CyclicalIterator(List<T>, T), the
iteration would start from the index of T plus one, leading to an
IndexOutOfBounds exception if T was the last element of the list.

To solve the problem use the convention that Java's iterator use: the
given object will be the element returned by the following call to
next(). Update MotherNature to reflect this change.

This also fixes a spurious ci job failure that at first seemed
nonsensical, however it made very much sense.