Client: implement connectivity checking
The client, to test the server's reachability, will periodically send
a HEARTBEAT message that the server will simply echo back. This will
keep the connection alive, allowing for a read timeout to be set.

The default intervals are 10s for the read timeout and 5s for the
interval between checks.

The protocol documentation will be updated in a followup commit.